integrate ai part
