+++ extension/_locales/ja/messages.json
